requirements.txt
pytorch
torchtext
transformers
editdistance
termcolor
tensorboardX
infomap
bcubed
progress
